requirements.txt
google_api_python_client==2.94.0
isodate==0.6.1
openai==0.27.8
pyairtable==1.5.0
python-dotenv==1.0.0
